Fix: removing malloc.h and including stdlib.h
C99 standard does not even mention `malloc.h`. `malloc` should be
declared in `stdlib.h` according to the standard.

`malloc.h` is no longer in the list of C Standard Library header
files.

With this change we can can build in the macOS (it does not ship
`malloc.h` header as part of C Standard lib).